tidied up item post code
