+++ save_craigslist_ui/lib/server_url.dart
@@ -1,5 +1,5 @@
-//const hostURL = 'http://localhost';
-const hostURL = 'http://10.0.2.2';
+const hostURL = 'http://localhost';
+//const hostURL = 'http://10.0.2.2';
 //const hostURL = 'http://benwichser.com';
 //const hostURL = 'http://172.16.1.166';
