+++ server/routes/items.js
@@ -30,52 +30,6 @@ function makeListingsOutput(obj) {
 }
 
 
-// Functions
-function itemPostTagEnhancer(body) {
-  /* itemPostTagEnhancer
-  * Scans title for useable tags, and adds them to body's tag array
-  * Accepts:
-  *   body (Object):  Object from item post
-  * Returns:
-  *   Null.  Alters `body['tags']`.
-  */
- // make new array based on any tags already there
-  let improvedTags = body.hasOwnProperty('tags') ? body['tags'] : [];
-  // remove punctuation from both title and tags, in that order -- 
-  // Not entirely sure ALL punctuation removal is best
-  // var punctuation = '!"#$%&\'()*+,-./:;<=>?@[\\]^_`{|}~';
-  var regex = new RegExp(/[\u2000-\u206F\u2E00-\u2E7F\\'!"#$%&()*+,\-.\/:;<=>?@\[\]^_`{|}~]/, 'g');
-  const noPunctuationTitle = body['title'].replace(regex, ' ');
-  // any entered tags are made to be a string and then reformed into list
-  //  after punctuation is removed.  This is to make sure the list is only
-  //  of single words
-  improvedTags = improvedTags.join(' ').replace(regex, ' ').split(' ');
-  // change all current tag words to lower case
-  if (improvedTags.length > 0)
-    improvedTags.forEach( (name, index) => improvedTags[index] = name.toLowerCase());
-  // add every title word that isn't alreadya tag into tag array
-  const titleArray = noPunctuationTitle.split(' ');
-  for (let word of titleArray)
-  {
-    if (!improvedTags.includes(word.toLowerCase()))
-    {
-      improvedTags.push(word.toLowerCase());
-    }
-  }
-  // remove common words, as well as trailing 's' and 't' from contractions/possessive
-  improvedTags = stopwords.removeStopwords(improvedTags, [...stopwords.en, ...['s', 't']]);
-  // quick convert to set and back to remove duplicates
-  improvedTags = [ ...new Set(improvedTags)];
-  // remove empty strings
-  improvedTags = improvedTags.filter( function(ele) {
-    return  ele != '';
-  }
-  );
-  // fix the body (ody ody ody) 'tags' value
-  body['tags'] = improvedTags;
-}
-
-// Router routes
 router.use(bodyParser.json());
 
 router.post(
@@ -91,7 +45,7 @@ router.post(
   //customValidation.isLoggedIn,
   customValidation.validate,
   (req, res) => {
-    itemPostTagEnhancer(req.body);
+    db.itemPostTagEnhancer(req.body);
     let new_id = uuidv4();
     let new_item = _.extend(req.body, { id: new_id });
     db.createItem("items", aws.DynamoDB.Converter.marshall(new_item));
@@ -100,23 +54,13 @@ router.post(
 );
 
 router.get("/", async (req, res) => {
+  // set default location to home of PBS's "Zoom" if there is none already given
   var zip = 'location' in req.body ? String(req.body.location) : '02134';
   try {
-    let listings = await db.getOpeningItemList(zip, 10);
+    let listings = await db.getItemList(req.body);
     res.status(201).json(makeListingsOutput(listings));
   } catch (err) {
-    console.log(`Error getting opening item list: ${err}`);
-  }
-});
-
-
-router.get("/search", async(req, res) => {
-  // Route for search from user
-  try {
-    let listings = await db.getSearchItems(req.body);
-    res.status(201).json(makeListingsOutput(listings));
-  } catch (err) {
-    console.log(`Error getting search item list: ${err}`);
+    console.log(`Error getting item list: ${err}`);
   }
 });
